+++ client/src/App.js
@@ -6,7 +6,6 @@ import Navbar from "./Components/Profile/Navbar/Navbar";
 import { useUserContext } from "./hooks/UseUserContext";
 import { useState } from "react";
 import axios from "axios";
-import HeaderEditForm from "./Components/HeaderEditForm/HeaderEditForm";
 import classNames from "classnames";
 
 function App() {
