+++ client/src/Components/Profile/Profile.jsx
@@ -37,25 +37,24 @@ export default function Profile() {
     return (
         <>
             <div className='div-container'>
-                {/* <ProfileOverview details={details} /> */}
-                <ProfileOverview />
+                <ProfileOverview details={details} />
                 <div className="row gx-1 container-row">
                     {/**Column for the left side (portfolio, experience etc) */}
                     <div className="col-8 ">
                         {/**That is 1 part (experience) until the next comment */}
-                        {/* <Experience experience={experience} /> */}
+                        <Experience experience={experience} />
                         {/**Next part starts here */}
-                        {/* <Portfolio /> */}
+                        <Portfolio />
 
                     </div>
                     {/**The column for the right side(qualification, certification etc) */}
                     <div className="col-3 r-col">
                         {/**Certification part of right column */}
-                        {/* <Certification />
+                        <Certification />
                         <Qualification />
                         <KeyTools />
                         <Interests />
-                        <Languages /> */}
+                        <Languages />
                     </div>
                 </div>
             </div>
