+++ client/src/Components/Profile/ProfileOverview/ProfileOverview.jsx
@@ -3,7 +3,6 @@ import "./ProfileOverview.css"
 import HeaderForm from "../../HeaderForm/HeaderForm"
 import { useUserContext } from '../../../hooks/UseUserContext';
 import axios from 'axios';
-import HeaderEditForm from '../../HeaderEditForm/HeaderEditForm';
 import { retrieveUser } from '../../../Utils/GetDetails.js'
 
 export default function ProfileOverview() {
@@ -11,46 +10,29 @@ export default function ProfileOverview() {
   const { user } = useUserContext();
 
   const [profileForm, setProfileForm] = useState(false);
-  const [overview, setOverview] = useState([]);
+  const [overview, setOverview] = useState({
+    firstName: '',
+    lastName: '',
+    nationality: '',
+    linkedInURL: '',
+    pronouns: '',
+    githubURL: '',
+    profileHeadline: '',
+    profileVideoURL: '',
+    profileImageURL: '',
+    tagline: ''
+  });
 
   const fetchUserData = async () => {
     const data = await retrieveUser(user);
-    setOverview(data.data[0]);
+    if (data.data) setOverview(data.data);  // what *should* be in use
+    // if (data.data[0]) setOverview(data.data[0]);  // temp because of backend code
   };
 
   useEffect(() => {
-
     fetchUserData();
-
-    // const getOverview = async () => {
-    //   const uri = `http://localhost:9000/profile`
-
-    //   const response = await axios.get(uri, {
-    //     headers: {
-
-    //       'Authorization': `Bearer ${user.token}`
-    //     }
-    //   })
-    //   setOverview(response.data[0])
-
-
-    // }
-    // getOverview();
-
-    // setOverview(details);
-
   }, []);
 
-  // const { profileImageURL,
-  //   firstName,
-  //   lastName,
-  //   nationality,
-  //   linkedInURL,
-  //   pronouns,
-  //   githubURL,
-  //   profileHeadline,
-  //   profileVideoURL,
-  //   tagline } = overview;
 
   return (
     <>
@@ -75,8 +57,7 @@ export default function ProfileOverview() {
 
             <h2>{overview.firstName + " " + overview.lastName}</h2>
             <button className='mod-btn' onClick={() => setProfileForm(true)}>
-              {profileForm && <HeaderEditForm overview={overview} />}
-              {/* {profileForm && <HeaderForm />} */}
+              {profileForm && <HeaderForm overview={overview} />}
               <img src='https://cdn-icons-png.flaticon.com/512/1159/1159633.png' />
             </button>
           </div>
