+++ client/src/Utils/GetDetails.js
@@ -13,13 +13,6 @@ export const retrieveUser = async (user) => {
   catch (e) {
     return { error: 'Problem getting user overview' }
   }
-
-  // try {
-  //   const response = await axios.get(`http://localhost:9000/profile/${user._id}`);
-  //   return response;
-  // } catch (e) {
-  //   return { error: "Error" };
-  // }
 };
 export const retrieveUserExperience = async () => {
   // try {
